% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPieDonut.R
\docType{data}
\name{browsers}
\alias{browsers}
\title{Browser market share 2011}
\format{An object of class \code{data.frame} with 11 rows and 3 columns.}
\usage{
browsers
}
\description{
A phony dataset measuring browser market share
}
\details{
@format A data.frame with 12 rows and 3 columns
 \describe{
    \item{browser}{browser}
    \item{version}{browser version}
    \item{share}{market share, in percentage}
}
}
\keyword{datasets}

