% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{get_popdata}
\alias{get_popdata}
\title{Grabs the required population data from the US Census Bureau’s International Data Base and outputs a data frame in the right format for ggplot2.}
\usage{
get_popdata(country, year)
}
\arguments{
\item{country}{\href{https://en.wikipedia.org/wiki/List_of_FIPS_country_codes}{a FIPS 10-4 country code} for the country you want}

\item{year}{The year you want to visualize.}
}
\value{
a data.frame
}
\description{
Modified from the source: rpubs.com/walkerke/pyramids_ggplot2
}
\examples{
#KS2016=get_popdata("KS",2016)
#ggBidirectionalBar(data=KS2016,left="Male",right="Female",label="Age")
#ggBidirectionalBar(data=KS2016,left="Male",right="Female",label="Age",interactive=TRUE)
}

