% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggAncova.default}
\alias{ggAncova.default}
\title{Make an interactive plot for an ANCOVA model}
\usage{
\method{ggAncova}{default}(data, y, x, A, label = NULL, digits = 1,
  interactive = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{y}{A character string of "continuous" column name be assigned to a response variable.}

\item{x}{A character string of "continuous" column name be assigned to a covariate.}

\item{A}{A character string of column name be assigned to a grouping variable.}

\item{label}{A character string of column name be assigned to the label}

\item{digits}{integer indicating the number of decimal places}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\description{
Make an interactive plot for an ANCOVA model
}

