% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggAncova.formula}
\alias{ggAncova.formula}
\title{Make an interactive plot for an ANCOVA model}
\usage{
\method{ggAncova}{formula}(formula, data, label = NULL, digits = 1,
  interactive = FALSE)
}
\arguments{
\item{formula}{A formula}

\item{data}{a data.frame}

\item{label}{A character string of column name be assigned to the label}

\item{digits}{integer indicating the number of decimal places}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\description{
Make an interactive plot for an ANCOVA model
}

