% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBar.R
\name{ggBar}
\alias{ggBar}
\title{Draw an interactive barplot}
\usage{
ggBar(data, xvar, fillvar, yvar = NULL, stat = "count",
  position = "stack", palette = NULL, width = NULL, digits = 1,
  horizontal = FALSE, yangle = 0, addlabel = FALSE, polar = FALSE,
  interactive = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{xvar}{A character string of column name be assigned to the x-axis variable}

\item{fillvar}{A character string of column name be assigned to the fill variable}

\item{yvar}{A character string of column name be assigned to the y-axis variable. Uses only when stat="identity".}

\item{stat}{The statistical transformation to use on the data for this layer, as a string
c("count","identity")}

\item{position}{Position adjustment. One of the c("fill","stack","dodge")}

\item{palette}{A character string indicating the color palette}

\item{width}{Bar width}

\item{digits}{integer indicating the number of decimal places}

\item{horizontal}{A logical value. If TRUE,a horizontal bar plot will be returned}

\item{yangle}{A integer. The value will be used adjust the angle of axis.text.y}

\item{addlabel}{A logical value. If TRUE, label will be added to the plot}

\item{polar}{A logical value. If TRUE, coord_polar() function will be added}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{...}{other arguments passed on to geom_bar_Interactive.}
}
\value{
An interactive barplot
}
\description{
Draw an interactive barplot
}
\examples{
require(moonBook)
require(ggplot2)
require(ggiraph)
ggBar(acs,"Dx","smoking",interactive=TRUE,width=1,colour="white",size=0.2,polar=TRUE)
ggBar(acs,"Dx","smoking",position="fill",addlabel=TRUE,horizontal=TRUE,width=0.5)
ggBar(acs,"Dx","smoking",position="fill",interactive=TRUE,addlabel=TRUE)
ggBar(acs,"Dx","smoking",position="dodge",interactive=TRUE)
ggBar(rose,Month,group,"value",stat="identity",polar=TRUE,palette="Reds",width=1,
      color="black",size=0.1,interactive=TRUE)
}

