% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{ggBidirectionalBar}
\alias{ggBidirectionalBar}
\title{Draw an interactive bidirectional bar plot}
\usage{
ggBidirectionalBar(data, left = NULL, right = NULL, label = NULL,
  title = "", legendposition = 1, interactive = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{left}{A character vector of column names be assigned to left-sided bar}

\item{right}{A character vector of column names be assigned to right-sided bar}

\item{label}{A character vector of column names be assigned to the label}

\item{title}{Plot title}

\item{legendposition}{An integer specifying the legend position. One of among c(1,2,3,4). Default value is 1.}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\value{
An interactive interactive bidirectional bar plot
}
\description{
Draw an interactive bidirectional bar plot
}
\examples{
require(ggplot2)
KS2016=get_popdata("KS",2016)
ggBidirectionalBar(data=KS2016,left="Male",right="Female",label="Age")
ggBidirectionalBar(data=KS2016,left="Male",right="Female",label="Age",legendposition=2)
ggBidirectionalBar(data=KS2016,left="Male",right="Female",label="Age",interactive=TRUE)
}

