% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{ggBidirectionalBar2}
\alias{ggBidirectionalBar2}
\title{Draw a static bidirectional bar plot}
\usage{
ggBidirectionalBar2(data, left = NULL, right = NULL, label = NULL,
  width = c(0.46, 0.55), title = "")
}
\arguments{
\item{data}{A data.frame}

\item{left}{A character vector of column names be assigned to left-sided bar}

\item{right}{A character vector of column names be assigned to right-sided bar}

\item{label}{A character vector of column names be assigned to the label}

\item{width}{An integer vector specifying the relative width of two barplots.}
}
\value{
A static bidirectional bar plot
}
\description{
Draw a static bidirectional bar plot
}
\examples{
require(grid)
require(ggplot2)
KS2016=get_popdata("KS",2016)
ggBidirectionalBar2(data=KS2016,left="Male",right="Female",label="Age")
}

