% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBoxplot.R
\name{ggBoxplot}
\alias{ggBoxplot}
\title{Draw boxplots of a data.frame}
\usage{
ggBoxplot(x, rescale = FALSE, horizontal = FALSE, interactive = FALSE,
  ...)
}
\arguments{
\item{x}{a data.frame}

\item{rescale}{if true, rescale the data.frame}

\item{horizontal}{if true, horizontal boxplots will be made}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{...}{other arguments passed on to geom_boxplot_interactive}
}
\description{
Draw boxplots of a data.frame
}
\examples{
ggBoxplot(mtcars)
ggBoxplot(mtcars,horizontal=TRUE,interactive=TRUE)
}

