% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggCatepillar}
\alias{ggCatepillar}
\alias{ggCatepillar.default}
\alias{ggCatepillar.formula}
\title{Make an interactive catepillar plot}
\usage{
ggCatepillar(x, ...)

\method{ggCatepillar}{formula}(formula, data, interactive = FALSE,
  digits = 1)

\method{ggCatepillar}{default}(df, y, x, group = NULL, interactive = FALSE,
  digits = 1)
}
\arguments{
\item{x}{An R object to ggCatepillar}

\item{formula}{A formula of type y ~ x + A}

\item{data}{A data}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{digits}{An integer indicating the number of decimal places}

\item{df}{A data.frame}

\item{A}{A character string of "numeric" column name be used as a y-axis variable}

\item{B}{A character string of column name be used as a grouping variable. Default value os NULL}

\item{C}{A character string of column name be used as a x-axis variable}
}
\value{
Ain interactive catepillar plot
}
\description{
Make an interactive catepillar plot
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Make an interactive catepillar plot

\item \code{default}: Make an interactive catepillar plot
}}
\examples{

}

