% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggCatepillarPlot}
\alias{ggCatepillarPlot}
\title{Make an interactive catepillar plot}
\usage{
ggCatepillarPlot(df, A, B = NULL, C, interactive = FALSE, digits = 1)
}
\arguments{
\item{df}{A data.frame}

\item{A}{A character string of "numeric" column name be used as a y-axis variable}

\item{B}{A character string of column name be used as a grouping variable. Default value os NULL}

\item{C}{A character string of column name be used as a x-axis variable}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{digits}{An integer indicating the number of decimal places}
}
\value{
Ain interactive catepillar plot
}
\description{
Make an interactive catepillar plot
}

