% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\name{ggChoropleth}
\alias{ggChoropleth}
\title{Draw an interactive choropleth map}
\usage{
ggChoropleth(data, map, fillvar = "총인구_명", colors = c("white",
  "orange", "red"), map_id = "code", tooltip = NULL, facetvar = NULL,
  subarea = NULL, title = "", interactive = FALSE, ...)
}
\arguments{
\item{data}{a data.frame}

\item{map}{a map maybe a result of map_data()}

\item{fillvar}{a column name assigned to a fill variable}

\item{colors}{A vector of colours used as a parameter of scale_fill_gradientn()}

\item{map_id}{a column name used as an id}

\item{tooltip}{a column name included in a tooltip}

\item{facetvar}{a column name assigned to a facet variable}

\item{subarea}{a name of subarea}

\item{title}{A title}

\item{interactive}{Logical. If positive an interactive map will be made}

\item{...}{other arguments passed on to geom_map_interactive}
}
\description{
Draw an interactive choropleth map
}
\examples{
crimes <- data.frame(state = tolower(rownames(USArrests)), USArrests)
require(ggplot2)
states_map <- map_data("state")
ggChoropleth(crimes,states_map,fill="Murder",map_id="state",tooltip="state",interactive=TRUE)
#ggChoropleth(data,kormap1,fillvar="총인구_명",tooltip="name",interactive=TRUE)
#ggChoropleth(data2,kormap2,fillvar="총인구_명",tooltip="name",interactive=TRUE)
#ggChoropleth(data3,kormap3,fillvar="총인구_명",tooltip="name",interactive=TRUE)
#ggChoropleth(data3,kormap3,fillvar="총인구_명",subarea=c("전라","광주"),interactive=TRUE)
}

