% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCLE.R
\name{ggCLE}
\alias{ggCLE}
\title{Draw a cleveland dot plot}
\usage{
ggCLE(data, xvar, yvar = NULL, colorvar = NULL, facetvar = NULL,
  reorderByX = TRUE, no = NULL, start = 0.99, interactive = FALSE,
  decreasing = TRUE, ...)
}
\arguments{
\item{data}{a data.frame}

\item{xvar}{A character string of column name be used as a x-axis variable}

\item{yvar}{A character string of column name be used as a y-axis variable.}

\item{colorvar}{A character string of column name be used as a colour variable. Default value is NULL}

\item{facetvar}{A character string of column name be used as a facetting variable. Default value is NULL}

\item{reorderByX}{If true, the data is reordered by x variable}

\item{no}{Number of data be drawn in plot}

\item{start}{start point of x axis as ratio to minimum x variable}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{decreasing}{Should the sort order be increasing or decreasing?}

\item{...}{other arguments passed on to geom_point}
}
\description{
Draw a cleveland dot plot
}

