% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCor.R
\name{ggCor}
\alias{ggCor}
\title{Draw a heatmap of correlation test}
\usage{
ggCor(df, label = FALSE, colors = NULL, title = FALSE,
  interactive = FALSE, ...)
}
\arguments{
\item{df}{A data.frame}

\item{label}{if true, add label to the heatmap}

\item{colors}{colors for low, mid and high correlation values}

\item{title}{if true, add title to the heatmap}

\item{...}{further arguments to be passed to cor.test}
}
\description{
Draw a heatmap of correlation test
}
\examples{
# ggCor(iris)
# ggCor(mtcars,interactive=TRUE)
# ggCor(iris,method="pearson")
}

