% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPieDonut.R
\name{ggDonut}
\alias{ggDonut}
\title{Draw a Donut plot}
\usage{
ggDonut(data = acs, donuts = "Dx", count = NULL, addDonutLabel = TRUE,
  showRatio = TRUE, polar = TRUE, labelposition = 1, title = "",
  interactive = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{donuts}{A character string of column name be assigned to the Donuts}

\item{count}{A character string of column name uses as count}

\item{addDonutLabel}{A logical value. If TRUE, labels are added to the Donuts}

\item{showRatio}{A logical value. If TRUE, Ratios are added to the DonutLabels}

\item{polar}{A logical value. If TRUE, coord_polar() function will be added}

\item{labelposition}{A number indicating the label position}

\item{title}{Plot title}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\value{
An interactive Pie and Donut plot
}
\description{
Draw a Donut plot
}

