% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDot.R
\name{ggDot}
\alias{ggDot}
\title{Draw a Wilkinson dot plot}
\usage{
ggDot(data, xvar, yvar = NULL, fillvar = NULL, stackdir = "center",
  binaxis = "y", binwidth = 0.5, method = "dotdensity", position = 0.2,
  boxwidth = 0.25, boxfill = NULL, ...)
}
\arguments{
\item{data}{a data.frame}

\item{xvar}{A character string of column name be used as a x-axis variable}

\item{yvar}{A character string of column name be used as a y-axis variable.}

\item{fillvar}{A character string of column name be used as a fill variable. Default value is NULL}

\item{stackdir}{which direction to stack the dots. "up" (default), "down", "center", "centerwhole" (centered, but with dots aligned)}

\item{binaxis}{The axis to bin along, "x" (default) or "y"}

\item{binwidth}{When method is "dotdensity", this specifies maximum bin width. When method is "histodot", this specifies bin width. Defaults to 1/30 of the range of the data}

\item{method}{"dotdensity" (default) for dot-density binning, or "histodot" for fixed bin widths (like stat_bin)}

\item{position}{Position adjustment. If 0, no adjustment.}

\item{boxwidth}{The width of boxplot}

\item{boxfill}{Fill color of boxplot}

\item{...}{other arguments passed on to geom_dotplot}
}
\description{
Draw a Wilkinson dot plot
}
\examples{
require(ggplot2)
require(gcookbook) # for data heightweight
require(moonBook) #for use data radial
ggDot(radial,yvar="height",xvar="sex",fillvar="sex",boxfill="white",position=0,binwidth=1,boxwidth=1)
ggDot(heightweight,yvar="heightIn",xvar="sex",binwidth=0.4)
ggDot(radial,xvar="height",fillvar="sex",binwidth=1)
ggDot(heightweight,xvar="heightIn")
ggDot(heightweight,xvar="heightIn",fillvar="sex")
}

