% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggEffect.lm}
\alias{ggEffect.lm}
\title{Visualize the effect of interaction between two continuous independent variables on a response variable}
\usage{
\method{ggEffect}{lm}(x, no = 1, probs = c(0.1, 0.5, 0.9), point = TRUE,
  xvalue = NULL, digits = 2, use.rownames = FALSE, interactive = FALSE)
}
\arguments{
\item{x}{An integer(1 or 2) indicating which independent variable is used as x-axis variable}

\item{probs}{A vector of probability weights for obtaining the elements of the vector being sampled.Default value is c(0.10,0.5,0.90)}

\item{point}{A logical value. If TRUE, draw points}

\item{xvalue}{A numeric vector}

\item{digits}{An integer indicating the number of decimal places}

\item{use.rownames}{If TRUE, use rownames in label}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\value{
An interactive plot showing interaction
}
\description{
Visualize the effect of interaction between two continuous independent variables on a response variable
}
\examples{
data(mtcars)
fit=lm(mpg~wt*hp,data=mtcars)
require(ggplot2)
ggEffect(fit,use.rownames=TRUE)
ggEffect(fit,use.rownames=TRUE,interactive=TRUE)
ggEffect(fit,no=2)
require(moonBook)
fit2=lm(NTAV~age*smoking,data=radial)
ggEffect(fit2,interactive=TRUE)
fit3=lm(age~sex*smoking,data=acs)
ggEffect(fit3,interactive=TRUE)
}

