% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggEffect}
\alias{ggEffect}
\alias{ggEffect.default}
\alias{ggEffect.formula}
\title{ggEffect generic}
\usage{
ggEffect(x, ...)

\method{ggEffect}{default}(x, y, x1, x2, ...)

\method{ggEffect}{formula}(x, data, ...)
}
\arguments{
\item{x}{Object to ggEffect}

\item{...}{additional arguments passed to the generic function}

\item{x1}{A name of one of the independent variable}

\item{x2}{A name of the other independent variable}

\item{data}{A data}

\item{x}{A name of response variable}

\item{x}{A formula}
}
\value{
An interactive plot showing interaction
}
\description{
Visualize the effect of interaction between two continuous independent variables on a response variable
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Visualize the effect of interaction between two continuous independent variables on a response variable

\item \code{formula}: Visualize the effect of interaction between two continuous independent variables on a response variable
}}
\examples{
#data(mtcars)
#fit=lm(mpg~wt*hp,data=mtcars)
#ggEffect(fit,use.rownames=TRUE)
#ggEffect(fit,no=2)
#require(moonBook)
#fit2=lm(NTAV~age*smoking,data=radial)
#ggEffect(fit2)
#fit3=lm(age~sex*smoking,data=acs)
#ggEffect(fit3,interactive=TRUE)
}

