% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggErrorBar.R
\name{ggErrorBar}
\alias{ggErrorBar}
\alias{ggErrorBar.default}
\title{Make an interactive bar plot with Errorbar}
\usage{
ggErrorBar(x, ...)

\method{ggErrorBar}{default}(x, yvar, xvar, group = NULL,
  interactive = FALSE, digits = 1, mode = 2)
}
\arguments{
\item{x}{An R object to ggErrorBar}

\item{yvar}{A character string of "numeric" column name be used as a y-axis variable}

\item{xvar}{A character string of column name be used as a grouping variable. Default value os NULL}

\item{group}{A character string of column name be used as a x-axis variable}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{digits}{An integer indicating the number of decimal places}

\item{mode}{if 2, two-sided error bar will be displayed, if 1 one-sided errorbar will be displayed}

\item{x}{A data.frame}
}
\value{
An interactive catepillar plot
}
\description{
Make an interactive bar plot with Errorbar
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Make an interactive bar plot with error bar
}}

