% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLogReg.R
\name{ggLogReg}
\alias{ggLogReg}
\title{Draw a jittered scatter plot for logistic regression model}
\usage{
ggLogReg(data, yvar, xvar, se = TRUE, interactive = FALSE, digits = 2,
  ...)
}
\arguments{
\item{data}{a data.frame}

\item{yvar}{A character string of column name be used as a response(y-axis) variable.}

\item{xvar}{A character string of column name be used as a explanatory(x-axis) variable}

\item{se}{Logical. display confidence interval around linear regression? (TRUE by default)}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{...}{other arguments passed on to geom_point}
}
\description{
Draw a jittered scatter plot for logistic regression model
}

