% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPair.R
\name{ggPairLong}
\alias{ggPairLong}
\title{Make an interactive scatter and line plot with long form data.frame}
\usage{
ggPairLong(data, idvar = "id", variablevar = "variable",
  valuevar = "value", colorvar = NULL, idcolor = TRUE,
  horizontal = FALSE, interactive = FALSE)
}
\arguments{
\item{data}{a data.frame with a long form}

\item{idvar}{a name of column used as an id variable}

\item{variablevar}{a name of column used as an variable}

\item{valuevar}{a name of column containa a value}

\item{colorvar}{names of column which is assigned as a colour variable}

\item{idcolor}{Logical cvalue. If TRUE, row numbers uses as a color variable}

\item{horizontal}{Logical cvalue. If TRUE, coord_flip() function is used to make a horizontal plot}

\item{interactive}{Logical cvalue. If TRUE, an interactive plot using ggiraph() function will be returned}
}
\description{
Make an interactive scatter and line plot with long form data.frame
}
\examples{
require(reshape2)
require(ggplot2)
iris$id=1:nrow(iris)
irislong=melt(iris[3:6],id=c("id","Species"))

ggPairLong(irislong,idvar="id",colorvar = "Species")
ggPairLong(irislong,idvar="id",colorvar = "Species",interactive=TRUE)
ggPairLong(irislong,idvar="id",interactive=TRUE)
}

