% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPieDonut.R
\name{ggPieDonut}
\alias{ggPieDonut}
\title{Draw a Pie and Donut plot}
\usage{
ggPieDonut(data = acs, pies = "Dx", donuts = "smoking", count = NULL,
  addPieLabel = TRUE, addDonutLabel = TRUE, showRatioDonut = TRUE,
  showRatioPie = TRUE, showRatioPieAbove10 = TRUE, title = "",
  labelposition = 1, polar = TRUE, interactive = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{pies}{A character string of column name be assigned to the Pies}

\item{donuts}{A character string of column name be assigned to the Donuts}

\item{count}{A character string of column name uses as count}

\item{addPieLabel}{A logical value. If TRUE, labels are added to the Pies}

\item{addDonutLabel}{A logical value. If TRUE, labels are added to the Donuts}

\item{showRatioDonut}{A logical value. If TRUE, Ratios are added to the DonutLabels}

\item{showRatioPie}{A logical value. If TRUE, Ratios are added to the PieLabels}

\item{showRatioPieAbove10}{A logical value. If TRUE, labels are added to the Pies with ratio above 10.}

\item{title}{Plot title}

\item{labelposition}{A number indicating the label position}

\item{polar}{A logical value. If TRUE, coord_polar() function will be added}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\value{
An interactive Pie and Donut plot
}
\description{
Draw a Pie and Donut plot
}
\examples{
#ggPieDonut(mtcars,"am","cyl")
#p<-ggPieDonut(mtcars,"am","cyl")
#ggiraph(code=print(p))
#ggPieDonut(polar=FALSE)
#ggPieDonut(labelposition=0)
#ggPieDonut(interactive=TRUE)
#ggPieDonut(mtcars,"cyl","am")
#ggPieDonut(browsers,"browser","version","share", title="Browser market share 2011",interactive=TRUE)
}

