% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{ggPoints}
\alias{ggPoints}
\alias{ggPoints.default}
\alias{ggPoints.formula}
\title{Draw scatterplot}
\usage{
ggPoints(x, ...)

\method{ggPoints}{default}(x, xvar, yvar, colour = NULL, fill = NULL,
  smooth = TRUE, method = "auto", formula = y ~ x, se = TRUE,
  level = 0.95, jitter = FALSE, fullrange = FALSE, interactive = FALSE,
  ...)

\method{ggPoints}{formula}(x, data, ...)
}
\arguments{
\item{x}{An R object to ggPoints}

\item{...}{other arguments passed on to ggPoint.defalut}

\item{xvar}{A character string of column name be used as a x-axis variable}

\item{yvar}{A character string of column name be used as a y-axis variable.}

\item{colour}{A character string of column name be used as a colour variable. Default value is NULL}

\item{fill}{A character string of column name be used as a fill variable. Default value is NULL}

\item{smooth}{A logical value. If TRUE, a layer of geom_smooth() added}

\item{method}{smoothing method (function) to use, eg. lm, glm, gam, loess, rlm.
For datasets with n < 1000 default is \code{\link[stats]{loess}}.
For datasets with 1000 or more observations defaults to gam, see \code{\link[mgcv]{gam}} for more details.}

\item{formula}{formula formula to use in smoothing function, eg. y ~ x, y ~ poly(x, 2), y ~ log(x)}

\item{se}{display confidence interval around smooth? (TRUE by default, see level to control}

\item{level}{level of confidence interval to use (0.95 by default)}

\item{jitter}{A logical value. If TRUE, a layer of geom_jitter() used instead of geom_point()}

\item{fullrange}{should the fit span the full range of the plot, or just the data}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{data}{A data.frame}

\item{x}{A data.frame}

\item{form}{A formula of type y ~ x | B}
}
\value{
An interactive plot or ggplot
}
\description{
Draw scatterplot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Draw scatterplot

\item \code{formula}: Draw scatterplot
}}
\examples{
# require(ggplot2)
# require(moonBook)
# ggPoints(iris,xvar=Sepal.Length,yvar=Sepal.Width,colour="Species",interactive=TRUE)
# ggPoints(mtcars,yvar=mpg,xvar=wt,fill="factor(cyl)",pch=21,smooth=FALSE)
# ggPoints(acs,xvar=height,yvar=weight,colour="sex",interactive=TRUE,
#           position="jitter",method="lm",formula=y~poly(x,2))
#ggPoints.formula(weight~height|sex,data=radial,interactive=TRUE)
#ggPoints.formula(weight~height,data=radial)
}

