% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBar.R
\name{ggRose}
\alias{ggRose}
\title{Draw an interactive Rose plot}
\usage{
ggRose(data, xvar, fillvar, ..., palette = "Reds", color = "black",
  size = 0.1)
}
\arguments{
\item{data}{A data.frame}

\item{palette}{A character string indicating the color palette}

\item{size}{Bar size}

\item{colour}{Bar colour}
}
\value{
An interactive Rose plot
}
\description{
Draw an interactive Rose plot
}
\examples{
require(moonBook)
require(ggplot2)
require(ggiraph)
ggRose(rose,Month,group,"value",interactive=TRUE)
ggRose(acs,"Dx","smoking",interactive=TRUE)
ggRose(rose,Month,group,"value",interactive=TRUE)
ggBar(rose,Month,group,"value",stat="identity",polar=TRUE,palette="Reds",width=1,
      color="black",size=0.1,interactive=TRUE)
}

