% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpine.R
\name{ggSpine}
\alias{ggSpine}
\title{Draw an interactive spinogram}
\usage{
ggSpine(data = acs, fillvar = "smoking", xvar = "age", yvar = NULL,
  stat = "count", position = "fill", palette = "Blues",
  interactive = FALSE, polar = FALSE, width = NULL, digits = 1,
  colour = "black", size = 0.2, addlabel = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{fillvar}{A character string of column name be assigned to the fill variable}

\item{xvar}{A character string of column name be assigned to the x-axis variable}

\item{yvar}{A character string of column name be assigned to the y-axis variable. Uses only when stat="identity".}

\item{stat}{The statistical transformation to use on the data for this layer, as a string
c("count","identity")}

\item{position}{Position adjustment. One of the c("fill","stack","dodge")}

\item{palette}{A character string indicating the color palette}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{polar}{A logical value. If TRUE, coord_polar() function will be added}

\item{width}{Bar width}

\item{digits}{integer indicating the number of decimal places}

\item{colour}{Bar colour}

\item{size}{Bar size}

\item{addlabel}{A logical value. If TRUE, label will be added to the plot}

\item{...}{other arguments passed on to layer.}
}
\value{
An interactive spinogram
}
\description{
Draw an interactive spinogram
}
\examples{
#require(moonBook)
#ggSpine(data=acs,"smoking","age",width=1,color="black",size=0.2)
#ggSpine(position="stack",addlabel=TRUE)
#ggSpine(addlabel = TRUE)
#ggSpine(data=acs,"Dx","smoking",addlabel=TRUE)
#ggSpine(data=acs,"Dx","smoking",position="dodge",addlabel=TRUE)
#ggSpine(data=acs,"Dx","smoking",position="stack",addlabel=TRUE)

}

