% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{human_gbp}
\alias{human_gbp}
\title{Human versions of large currency numbers - for "£"}
\usage{
human_gbp(x)
}
\arguments{
\item{x}{a numeric vector to format, smbl a symbol you'd like to prefix your numbers by}
}
\value{
a character vector the same length as the input vector
}
\description{
Human versions of large currency numbers - for "£"
}

