% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\docType{data}
\name{kormap3}
\alias{kormap3}
\title{Korean administative map data(2010) level 3}
\format{An object of class \code{data.frame} with 70711 rows and 11 columns.}
\usage{
kormap3
}
\description{
A dataset
}
\details{
@format A data.frame with 70711 rows and 11 columns
 \describe{
    \item{id}{Korean administative area code}
    \item{long}{longitude}
    \item{lat}{latitude}
    \item{order}{An integer}
    \item{hole}{A logical}
    \item{piece}{Factor with 11 levels}
    \item{group}{Factor with 3593 levels}
    \item{SP_ID}{shape id}
    \item{FID}{An integer}
    \item{code}{Area code}
    \item{region}{Region}
}
}
\keyword{datasets}

