% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{multiggplot}
\alias{multiggplot}
\title{Draw a multiggplot}
\usage{
multiggplot(p, vp, title)
}
\arguments{
\item{p}{A list of ggplot}

\item{vp}{A list of viewport}

\item{title}{Plot title}
}
\description{
Draw a multiggplot
}

