% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\name{paste_or}
\alias{paste_or}
\title{Paste '|' between vectors}
\usage{
paste_or(...)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}
}
\description{
Paste '|' between vectors
}

