% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{pastecolon}
\alias{pastecolon}
\title{Paste character vactoers with separatoe colon}
\usage{
pastecolon(...)
}
\arguments{
\item{...}{Arguments passed on to paste()}
}
\description{
Paste character vactoers with separatoe colon
}

