% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBidirectionalBar.R
\name{PopPyramid}
\alias{PopPyramid}
\title{Get population data and draw popuation pyramid}
\usage{
PopPyramid(country, year, interactive = FALSE)
}
\arguments{
\item{country}{\href{https://en.wikipedia.org/wiki/List_of_FIPS_country_codes}{a FIPS 10-4 country code} for the country you want}

\item{year}{The year you want to visualize.}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\value{
A population pyramid
}
\description{
Get population data and draw popuation pyramid
}
\examples{
#PopPyramid("KS",2016)
#PopPyramid("NI",2015)
#PopPyramid("JA",2015)
#PopPyramid("VQ",2015)
}

