% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratetable.kor.R
\docType{data}
\name{ratetable.kor}
\alias{ratetable.kor}
\title{ratetable to calculated expected survival rate of Korean}
\format{An object of class \code{ratetable} of dimension 101 x 2 x 18.}
\usage{
ratetable.kor
}
\description{
ratetable.kor
}
\details{
@format A ratetable for Korean population 1997-2014
}
\keyword{datasets}

