% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPieDonut.R
\docType{data}
\name{taco}
\alias{taco}
\title{Taco ratings by age group}
\format{An object of class \code{data.frame} with 136 rows and 4 columns.}
\usage{
taco
}
\description{
Taco ratings by ShellType, AgeGroup and Filling
source: \href{http://rikturr.com/blog/communicating-experimental-results-with-r/}{Communicating experiment results with R}
}
\keyword{datasets}

