% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggAncova}
\alias{ggAncova}
\title{Make an interactive plot for an ANCOVA model}
\usage{
ggAncova(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments passed to the generic function}
}
\description{
Make an interactive plot for an ANCOVA model
}

