% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggAncova}
\alias{ggAncova}
\alias{ggAncova.default}
\alias{ggAncova.formula}
\alias{ggAncova.lm}
\title{Make an interactive plot for an ANCOVA model}
\usage{
ggAncova(x, ...)

\method{ggAncova}{default}(data, y, x, A, label = NULL, digits = 1,
  interactive = FALSE)

\method{ggAncova}{formula}(formula, data, label = NULL, digits = 1,
  interactive = FALSE)

\method{ggAncova}{lm}(fit, label = NULL, digits = 1, interactive = FALSE)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments passed to the generic function}

\item{data}{a data.frame}

\item{y}{A character string of "continuous" column name be assigned to a response variable.}

\item{A}{A character string of column name be assigned to a grouping variable.}

\item{label}{A character string of column name be assigned to the label}

\item{digits}{integer indicating the number of decimal places}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{formula}{A formula}

\item{fit}{An object of class "lm"}

\item{x}{A character string of "continuous" column name be assigned to a covariate.}

\item{data}{a data.frame}

\item{label}{A character string of column name be assigned to the label}

\item{digits}{integer indicating the number of decimal places}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}
}
\description{
Make an interactive plot for an ANCOVA model
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Make an interactive plot for an ANCOVA model

\item \code{formula}: Make an interactive plot for an ANCOVA model

\item \code{lm}: Make an interactive plot for an ANCOVA model
}}

