% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAncova.R
\name{ggCatepillar}
\alias{ggCatepillar}
\alias{ggCatepillar.default}
\alias{ggCatepillar.formula}
\title{Make an interactive catepillar plot}
\usage{
ggCatepillar(x, ...)

\method{ggCatepillar}{formula}(x, data, ...)

\method{ggCatepillar}{default}(x, yvar, xvar, group = NULL,
  interactive = FALSE, digits = 1)
}
\arguments{
\item{x}{An R object to ggCatepillar}

\item{data}{A data}

\item{yvar}{A character string of "numeric" column name be used as a y-axis variable}

\item{xvar}{A character string of column name be used as a grouping variable. Default value os NULL}

\item{group}{A character string of column name be used as a x-axis variable}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{digits}{An integer indicating the number of decimal places}

\item{x}{A formula of type y ~ x + A}

\item{x}{A data.frame}
}
\value{
An interactive catepillar plot
}
\description{
Make an interactive catepillar plot
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Make an interactive catepillar plot

\item \code{default}: Make an interactive catepillar plot
}}

