% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggScatter.R
\name{ggScatter}
\alias{ggScatter}
\alias{ggScatter.default}
\alias{ggScatter.formula}
\title{Make an interactive scatterplot}
\usage{
ggScatter(x, ...)

\method{ggScatter}{default}(x, xvar, yvar, colorvar = NULL, se = TRUE,
  addloess = FALSE, loessse = FALSE, fullrange = FALSE, tooltip = NULL,
  interactive = FALSE)

\method{ggScatter}{formula}(x, data, ...)
}
\arguments{
\item{x}{An R object to ggPoints}

\item{...}{other arguments passed on to ggScatter.defalut}

\item{xvar}{A character string of column name be used as a x-axis variable}

\item{yvar}{A character string of column name be used as a y-axis variable.}

\item{colorvar}{A character string of column name be used as a colour variable. Default value is NULL}

\item{se}{Logical. display confidence interval around linear regression? (TRUE by default)}

\item{addloess}{Logical. add loess line}

\item{loessse}{Logical. display confidence interval around loess regression? (TRUE by default)}

\item{fullrange}{should the fit span the full range of the plot, or just the data}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{data}{a data.frame}

\item{x}{A data.frame}

\item{x}{A formula of type y ~ x | B}

\item{data}{A data.frame}
}
\description{
Make an interactive scatterplot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Draw scatterplot

\item \code{formula}: Draw scatterplot
}}
\examples{
require(moonBook)
require(ggplot2)
require(ggiraph)
ggScatter(iris,xvar="Sepal.Length",yvar="Sepal.Width",se=FALSE,colorvar="Species",interactive=TRUE)
ggScatter(acs,xvar="height",yvar="weight",colorvar="Dx",se=FALSE,interactive=TRUE)
ggScatter(iris,xvar="Sepal.Length",yvar="Sepal.Width",addloess=FALSE,colorvar="Species",interactive=TRUE)
ggScatter(radial,xvar="height",yvar="weight",interactive=TRUE)
require(moonBook)
ggScatter(weight~height|sex,data=radial,interactive=TRUE)
ggScatter(weight~height|sex,data=radial,fullrange=TRUE,se=FALSE,interactive=TRUE)
ggScatter(weight~height,data=radial,interactive=TRUE)
}

