% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggScatter.R
\name{ggScatter}
\alias{ggScatter}
\alias{ggScatter.default}
\alias{ggScatter.formula}
\title{Make an interactive scatterplot}
\usage{
ggScatter(x, ...)

\method{ggScatter}{default}(x, yvar, xvar, colorvar = NULL, fillvar = NULL,
  facetvar = NULL, se = TRUE, method = "auto", fullrange = FALSE,
  maxfactorno = 6, digits = 2, tooltip = NULL, interactive = FALSE, ...)

\method{ggScatter}{formula}(x, data, ...)
}
\arguments{
\item{x}{An R object to ggPoints}

\item{...}{other arguments passed on to geom_point}

\item{yvar}{A character string of column name be used as a y-axis variable.}

\item{xvar}{A character string of column name be used as a x-axis variable}

\item{colorvar}{A character string of column name be used as a colour variable. Default value is NULL}

\item{fillvar}{A character string of column name be used as a fill variable. Default value is NULL}

\item{facetvar}{A character string of column name be used as a facet variable. Default value is NULL}

\item{se}{Logical. display confidence interval around linear regression? (TRUE by default)}

\item{method}{smoothing method (function) to use, eg. "lm", "glm", "gam", "loess", "rlm"}

\item{fullrange}{should the fit span the full range of the plot, or just the data}

\item{tooltip}{A character string of column name be included in tooltip. Default value is NULL}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{data}{a data.frame}

\item{x}{A data.frame}

\item{x}{A formula of type y ~ x | B}

\item{data}{A data.frame}

\item{...}{other arguments passed on to ggScatter.defalut}
}
\description{
Make an interactive scatterplot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Draw scatterplot

\item \code{formula}: Draw scatterplot
}}
\examples{
require(moonBook)
require(ggplot2)
require(ggiraph)
ggScatter(iris,yvar="Sepal.Width",xvar="Sepal.Length",method="lm",interactive=TRUE)
ggScatter(iris,yvar="Sepal.Width",xvar="Sepal.Length",se=FALSE,colorvar="Species",interactive=TRUE)
ggScatter(acs,yvar="weight",xvar="height",colorvar="Dx",facetvar="Dx",se=FALSE,method="lm",interactive=TRUE)
ggScatter(iris,yvar="Sepal.Width",xvar="Sepal.Length",colorvar="Species",interactive=TRUE)
ggScatter(radial,yvar="NTAV",xvar="age",colorvar="DM",facetrvar="DM",interactive=TRUE)
require(moonBook)
ggScatter(weight~height|sex,data=radial,interactive=TRUE)
ggScatter(weight~height|sex,data=radial,fullrange=TRUE,se=FALSE,interactive=TRUE)
ggScatter(weight~height,data=radial,interactive=TRUE)
}

