% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{makeEquation}
\alias{makeEquation}
\title{Make a regression equation with regression model}
\usage{
makeEquation(fit, digits = 2)
}
\arguments{
\item{fit}{a regression model}

\item{digits}{An integer indicating the number of decimal places}
}
\value{
a regression equation
}
\description{
Make a regression equation with regression model
}

