% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\name{subdata}
\alias{subdata}
\title{Select subdata of data}
\usage{
subdata(data, area)
}
\arguments{
\item{area}{a string of area looking for}

\item{map}{an object of data.frame or class Shape(SpatialPolygonsDataFrame)}
}
\value{
Subdata of data.frame orclass Shape of which code matched with area
}
\description{
Select subdata of data
}

